Make several fields on "fixes" not nill

+++ src/device_repo.rs
@@ -96,12 +96,12 @@ impl DeviceRepository {
     /// Get a device model (with UUID) by address and address type
     pub async fn get_device_model_by_address(
         &self,
-        address: u32,
+        address: i32,
         address_type: AddressType,
     ) -> Result<Option<DeviceModel>> {
         let mut conn = self.get_connection()?;
         let device_model = devices::table
-            .filter(devices::address.eq(address as i32))
+            .filter(devices::address.eq(address))
             .filter(devices::address_type.eq(address_type))
             .first::<DeviceModel>(&mut conn)
             .optional()?;
